src/main/java/org/bigdatatechcir/learn_llm/model/requirements.txt
pytorch==2.4.0 
torchvision==0.19.0
datasets==2.16.1
jsonlines==4.0.0
marshmallow==3.22.0
matplotlib==3.5.1
ngrok==1.4.0
nltk==3.8
numpy==1.26.4
openai==1.42.0
pandas==1.5.3
psutil==5.9.8
pydantic==2.8.2
rich==13.7.1
scikit_learn==1.5.1
sentence_transformers==2.3.1
simhash==2.1.2
tiktoken==0.5.1
transformers==4.44.0
jinja2==3.1.2
jsonlines==4.0.0
trl==0.11.3
ujson==5.1.0
swanlab